requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
SQLAlchemy==2.0.30
alembic==1.13.1
Pillow==10.3.0
pillow-heif==0.16.0
python-docx==1.1.2
reportlab==4.2.0
fpdf2==2.7.9
LibreOffice          # system-level; installed via Dockerfile
python-dotenv==1.0.1
aiofiles==23.2.1
structlog==24.2.0
tenacity==8.3.0
psutil==5.9.8
humanize==4.9.0
